Yayy! Fixed the bug in my_p4est_nodes.c file.
Bug was due to a wrong update in the p.piggy3.local_num field for shared nodes. This was because in the previous implementation it was forgoten to update the nodes using old_indecies after sorting the list. Seems working now (check the main.cpp file in examples/parallel_vtk/ folder)