(Read detailed msg) Fixed the segfault coming from wrong ghost locidx and a bug in my_p4est_nodes.c file.
The code now runs as long as we wont have the issue with remote quadrant not being on the processor boundary. This will never be a problem if you are interpolating at backtraced poinst but _might_ cause issues when transferring results between grids and needs to be addressed anyways. Please take a look at the comments in bilinear_interpolation_function.cpp files.