Edited Poisson example to make it compatible with changes. Debugged vtk writer. (see details)
As of now, the mesh structure only seems to work if the grid is balanced, oterwise it breaks. However, the new node data structure (meaning treating hanging nodes as indep nodes) does not seem to affect the solver when the grid is balanced.

There was a bug in the cell writer of my_p4est_vtk.c file which caused problems when cell data other than proc_rank and tree_idx were added to the list. It is removed now.