Test for my_p4est_nodes_new with added ghost layer passed for all tests (see details)
As of now it is possible to pass ghost layer as an additional term to add nodes of ghost layer to the ghost nodes of p4est_nodes_t structure. If you do not wish to include this layer (a.k.a. old case) simply pass NULL as the second argument. All 4 tests currently work with the new structure.

Note, however, that interpolation class still assumes that the ghost layer is not included in the node structure. This obviously works but may be inefficient when ghost layer nodes are included in the node struture. We should change the class such that if this data is available, it would use the added info and prevent MPI for ghost cells.

Finally, I just wanted to add that now the my_p4est_vtk_write_all interface is changed to provide support for writing ghost layer if user wishes to. If you do not wish to write the ghost layer, simply pass NULL as the third argument. Also note that the scaling property is removed from the interface. The writer has been tested and works both in the new version and the old compatibility mode.