Updated SL/Interpolation to use new ghost structure (see details)
Now SL and (to some extent) interpolation classes use extra information from the nodes of ghost layer. Interpolation function further needs to be modified such that it wont require communication for pointsthat end up in the ghost layer.

Also, somehow the changes we made to the lookup_function (rounding, etc) now causes a weired bug in the SL when updating grids. It may or may not happen depending on the number of processors and the result depends on the epsilon we choose for perturbation in the said function. Definately needs revisiting