Modified interpolation function to use new ghost nodes (see details).
As of now, interpolation function will operate such that it will utilize information on the nodes of ghost layer. This means it is ABSOLUTELY necessary from now on to pass the ghost layer to 'my_p4est_nodes_new' when creating the nodes structure.

The rationale behond this is the most other parts, like Quadratic interpolation (to be done) and reinitialization already have ths requirement on the nodes structure so it makes perfect sense to add this for linear_interpolation as well.

Still having issues with the lookup_function though!