Added an efficient way of computing 2nd derivatives (see details)
three methods have been added to compute either dxx, or dyy, or both dxx and dyy. They all first only compute derivatives in the boundary region so that ghost update can be called immediately and before computing derivatives on interior points.

If one is interested in both dxx and dyy at the same time, which happens to be almost always teh case, one should call 'dxx_and_dyy_central' method instead as it uses a block vector. This cuts down the number of MPI messages by half (although 2x time message size) which should be more efficient for smaller problems since computation time is dominated by latency and not bandwidth. Note that this, however, comes at the cost of user convenience (see examples/derivatives)

This is as efficient as it can get!