cleaned up in my_p4est_levetset (see details)
Changed the update pattern of reinitialization to make it simpler and easier to understand. In the node_neighbors data structure two set of integer maps are defined
1) layer nodes: These are local nodes that are ghost to other processors
2) local nodes: These are local nodes that are not ghost to other processors

Any computation that requires updating on ghost values, e.g. derivatives, reinitialization, advection in normal direction etc., can now be done easily and more efficiently. This is a 4-step process:
1) do computation on the nodes mapped by layer_nodes
2) call VecGhostUpdateBegin to initiate ghost update message transfer
3) do computation on the nodes mapped by local_nodes
4) call VecGhostUpdateEnd to finialize the update process of ghost values

Since local_nodes.size() >> layer_nodes.size() for a moderatly big problem, one would hope that the amount of computation done in step 3 can sucessfully hide the latency and communication time associate with step 2 and 4.

Finally, a new macro is introduced, DXX_USE_BLOCKS. If this macro is defined by the compiler, then 'dxx_and_dyy_central' method uses a temporary block vector to calculate fxx and fyy and then copies the results back into fxx and fyy. Initial testing suggests that although this is reduces the number of messages by half, the final result is actually worse than calling the original function, probably due to losts of memory allocations, and to a lesser degree, extra copies. It may be possible to counter this effect by retaining the block vector at the cost increasing memory consumption. This is NOT implemented and postponed to future tests on Stampede.