motion in the normal direction and reinitialization work in 3D + debugging (See details)
So it turns out there was  a bad bug in my_p4est_node_neighbors class. Problem was related to the corrections when a node was on the wall. In these cases it is need to fetch second order neighbors in the opposite directions. Most of the time this just works fine. However, since we only have the first order information on ghost cells, it is possible for the function to return a wrong  cell that does not locally exist.

This happens if the first neighbor (from which the second neighbor is constructed) happens to be a ghost cell. In such a case there is no guarantee that the second neighbor (i.e. neighbor of the ghost cell) actually exists locally. To fix this only local first order neighbors are selected. Unfortunately this means that there is no guarantee that the second neighbor is actually the smallest. It is NOT possible to fix this problem without pest supporting more than single layer of ghost cells. That said, this could only happen for neighboring nodes on the wall and not interior nodes so the effect should really be minimal.