Fixed the bug causing the code to break (see details)
So it turn out that the bug was actually in the modification I made to the my_p4est_nodes.c during commit 31b38a33. I am still not 100% sure why it was causing the code to break but my best guess I was indexing the array out of its bound and that somehow caused corruption in the p4est object … very weird but possible. Nonetheless we should keep an eye on this file. To ensure proper behavior a check is added to the end of function.