Added the IDW interpolation method for cells ... it sucks! (see details)
Very simple to implement and is very cheap but creates really bad results at the T-junction. It may be possible to make it better if cells across corners (and edges in 3D) are also included in the cell neighbor structure ...

Next try is LSQR which is relatively easy to implement but very expensive! Implementation of linear interpolation in 3D looks like a nightmare (for now)!