Merged all changes from the 'run/stampede' branch (see details).
A couple of important changes have been introduced which I try to include here:
1) Interpolation class can now work using nonblocking collectives. This is included in MPI 3.0 standard which is relatively new
and might not work with all MPI libraries. This can be enabled by passing 'ENABLE_NONBLOCKING_NOTIFY' flag to the compiler. Also
note that on some systems (like stampede) the API is provided as an extension in earlier MPI libraries. For this to work, you also
need to pass 'ENABLE_MPI_EXTENSIONS' flags. The default behavior does not use nonblocking collectives.

2) From now on, 'my_p4est_node_neighbors_t' does not cache neighbors by default. If this is desired, user needs to manually call
appropriate 'init_neighbors()' function. Also the 'operator[]()' is intended to be used with initialized objects and will throw
otherwise. A new function 'get_neighbors(n, qnnn)' should be used if it is known that the neighbors are not initialized. Finally,
the function 'get_neighbors(n)' is designed to always work smartly regardless of whether or not the object is initialized.

3) 'my_p4est_vtk_write_all' can now automatically strip the leading folder name from a filename when writing pvtu files.

4) 'hierarchy' and 'my_p4est_node_neighbors_t' can now update themselves based on changes to the 'p4est', 'ghost', and 'nodes'

5) 'my_p4est_log_wrappers.cpp' was changed to 'my_p4est_log_wrappers.c'