updated introduction and added refs (see details)
currently newer refs are added in the refs.bib until we finialize on
which ones we actually want to retain. For the older ones we use the
master ref for CASL.

To cope with the fact that the file could be in different locations on
different machines, please add a symlink casl.bib to point to the actual
reference file on you machine. This file was not added to the repo since
the address could be different.

A better approach is to put the reference file in a system-provided
location that can be accessed by latex without the need for any relative
path.