changed purely Neumann case to use nullspace. Not working (see details)
The correct way of handling purely Neumann problem is to remove the
Nullspace from the matrix. However, since our discretization is non-
symmetric at the T-junctions, the constant vector is _NOT_ the correct
left nullspace although it is close. As a result, the solver may or may
not work and the results may be wrong

The other approach for handling these cases is to fix the solution at a
point. However, this approach gives us totally wrong results ... I guess
for now we should switch to using Dirichlet BC? ...