Fixed a minor bug when ghost == NULL. 3D is fine (see details)
So the issue of 3D was not related to 3D per se. The issue showed minor
defects at the processor boundary which were (more) visible in 3D. These
turn out to be due to overlapping of ghost cells with local ones when
ghost cells are also written and is visible both for periodic as well as
non-periodic domains.

Unless one is indeed interested in the ghost layers to be present inside
each .vtu file, I think it is best to avoid writing them due to these
*cosmetic* issues they create in the visualization. This is achieved by
passing NULL for the ghost layer in the vtk_write_all function.