Fix an issue when domain was periodic with only one tree (see details)
There was a problem with qnnn for nodes at the boundary when the domain
was periodic and only had a single tree in the periodic direction. A
single tree always maps to itself through calling tree_to_tree on the
faces. However, this is OK if the domain is periodic and should not be
handled any different from regular cases.