trying out levelset smoothing and two-time-scale integration (see details)
* By two-time-scale integration, we mean to integrate the advection many
small time-steps which are O(dx^3) to ensure stability while solving the
poisson equation every dt_large = O(dx). Implemented for godunov scheme
because it was easier but did not help with the stability problem

* Trying smoothing of levelset by solving a diffusion equation on phi
does get rid of stability issue but destroys solution features as well.
Does not seem to be easy to control.