A very serious bug in voronoi solver in parallel ... (see details)
* I am able to reproduce this bug with two_fluid_2d with the following
options: -pc_type sor -lip 2 -lmin 3 -M 0.01 and on 16 processors. The
code crashes on iteration 165 since the linear solver does not converge
, i.e. PETSc produces error message
Linear solve did not converge due to DIVERGED_PCSETUP_FAILED iterations 0
This in turn generates NAN in pressure field and leads to all sort of
hellish error messages which took me an entire day to trace back ...

Anyway, for whatever reason, it looks like the error is somehow linked
with the voronoi solver not being able to find a good partition which
then creates a zero on diagonal of linear system.

Now, SOR precoditioner does not work if there is a zero diagonal.
Interestingly hypre seems to be able to handle the case! For now I'm
ignoring this issue but it must definitely be resolved.