added some more measurements: applied net voltage, applied net Electric field, and fixed some bugs.
note: as density increases one has to increase the resolution to provide enough number of nodes in the vicinity of each pair of cells (become too close!) otherwise the solver crashes. I have seen crashes in the interpolation from Voronoi to tree specifically, it can not find neighboring points (I guess it need 4 points, but there aren't).

IMPORTANT:
The minimum resolution needed for a box of dimension 1mm is at least 8, otherwise not enough points to represent each cell.