Group meeting #1, preparation: illustrating data structures - Petsc
Basic usage of the PetSc library. A new example has been developed: it is totally independent of the rest of the library and intends to illustrate Petsc usage only. 
We play with calculations on a parallel uniform one-dimensional grid and illustrate basic features of Petsc usage: 
- how to create parallel vectors;
- how to access local and ghost values given an indexing convention;
- how to overlap non-blocking communications and local calculations;
- how to solve a simplistic linear system of equations;
- how to remap values from an origin parallel vector to a different destination parallel vector;
- how to export such vectors (along with a matlab script to read those exported vectors)

Some paths need to be changed to be consistent with the user's work environment in paths_and_compilers.pri as well as in my_script.m to be used right away