Changed and added bash scripts for easy configurations
Modified the existing newprofile bash script to make it consistent with Daniil's suggestion for easy interactions between developers and/or users. 
Every new environment must now be defined within a .pri file stored in the qmake directory. The bash script new_environment has been added to that directory to make the creation of such file easy and as error-proof as I could.