Working on refine and coarsen w fields, cleaning some interp updates
Refine and coarsen with a variable number of fields and user specified options is now working. In progress: update_p4est() function with variable number of fields, and option for additional LSF's to refine around. Currently, still seems to not be registering refinement options for the second field --> to be continued ... 

For interpolation --> got the new interpolation features implemented by Raphael, had to clean some functions in trajectory_of_point and in my_p4est_semi_lagrangian to work in accordance with these new tools