Amendments to merge conflicts, to make the library compile and work fine
Compiled and tested successfully (in DEBUG mode for elementary tests, hence passing all internal sanity checks) on my local workstation for:
- quad_neighbor_nodes_of_node, 2D/3D (execution and accuracy check);
- cell_xgfm, 2D/3D (execution and accuracy check);
- poisson_faces, 2D/3D (execution and accuracy check, including with a nonzero diagonal for dirichlet interface and neumann walls - common usage in Navier-Stokes);
- poisson_jump_voronoi, 2D/3D, with partition check  (execution and accuracy check). It works as well as before (which means it does !NOT! work in 3D with fine enough grids);
- interpolation_between_grids, 2D/3D, all combinations of methods (sequential, array of Vecs or block vectors) and all 5 interpolation procedures (linear, minmod, quadratic, Daniil_v1, Daniil_v2), with or without precomputed second derivtaives;
-  flow_past_sphere, 2D, lmin 4, lmax 8. The unconstrained interpolation of velocities at the outflow wall condition (Neumann boundary condition) posed issues. Therefore I decided to re-introduce it (more details in note 3 here below). Restart fixed and tested successfully, as well.
-  flow_past_sphere, 3D, lmin 4, lmax 6. Restart fixed tested successfully as well.
All runs and tests completed locally. Checks on Stampede2 to follow in the subsequent commits...

Some notes and insightful details:
1) my_p4est_levelset: this is nothing but my own OCD at work --> most modifications are nothing but making the code more readable... I was tracking a bug but it was actually living elsewhere. It looks like there is quite a lot of code duplication in those functions though, it could be good to avoid that and clean a bit if possible;
2) my_p4est_utils.cpp & my_p4est_interpolation_nodes : this is where the bug mentioned above was, in the way the periodic boundary conditions are handled --> the point could lie far away from the owning quadrant. First I alleviated code duplication in the basic interpolation routines in my_p4est_utils.cpp. Those routines now all assume that the Cartesian coordinates of the points are as close as possible to the provided quadrant (considering possible periodic wrapping). These routines are called from my_p4est_interpolation_nodes (only) and this class now handles that before passing the correct coordinates to the basic routines. A similar fix for periodic wrapping had been added to my_p4est_poisson_jump_voronoi.cpp.
3) my_p4est_solve_lsqr: there was some serious issue and/or code duplication in that. I got rid of all code duplication by making one very general function for each task, and calling it in different ways using wrapper functions (to keep the same interface).
solve_cholesky was supposed (and assumed) to return 'false' and abort the resolution in case of singular lsqr systems to solve. In fact, it would still solve systems apparently successfully in degenerate cases (very large condition number obtained numerically but still representable as a double number). 
This led to very seriously bad results and completely unreliable behavior or the lsqr interpolation --> I added a threshold value on the condition number for the choleski factor and set the default value of that threshold to be 1.0e4 (seems to alleviate the issue). 
In fact, a value as small as 10 or 20 can already lead to fairly spurious results apparently. This happened and was observed  when interpolating velocities from faces to Neumann WALL nodes without constraining parameters in the lsqr interpolant to match the given Neuman boundary condition. Therefore, I made the decision to constrain the lsqr interpolation in interpolate_f_at_node_n (renamed 'interpolate_velocity_at_node_n') in presence of Neumann boundary conditions (this is how it was implemented when wrapping up the parallel N-S paper, by the way).