IMPORTANT FIX FOR SEMI-LAGRANGIAN ADVECTION: TRAJECTORY OF POINTS
I most probably introduced the bug a while ago. I don't know when but this was is most probably on me: I'm very sorry!

Basically, the interpolation used to backtrace v or w faces was off : it used the velocity interpolated at u face of same local index. --> probably fairly similar, that's why it "seemed" to work alright but, still, it was wrong... Very sorry about that :-(