Non-blocking MPI send/recv do not like large tag values, apparently :-/
Everything was running fine on a single node but crashed on several because of this. This can't really be found in the MPI documentation (MPI_TAG_UB seems to be larger than what is actually allowed for real). I found the key online, I hope it works for good now. This is probably one of the stupidest bugs I had to kill, ever...