No longer relies on face intersections to trigger communications of CFs
--> any CF associated with a layer cell is now communicated to all relevant neighbor procs that may see it as a ghost cell.
Previously, the communication was done only if apparently relevant, i.e., if the shared face was seen as divided between the subdomains. However, it turns out that this subdivision on a shared face may differ when seen from one side or the other, in particular in 3D. 
Therefore it is not unequivocally defined and that may be responsible for an inconsistency-break within the previously used coding assumptions. Several recent 3D simulations crashed because of this: a given cell required the CF on a ghost cell but that ghost cell didn't see it as a relevant communication (face as seen from this side was small but nonzero, but was exactly 0.0 when seen from the ghost side). The CF was locally unknown, the proc was trying to access an inexistent member of a map, hence the crash...