Resolving issues with ml_mass_conservation project after long_overdue_merge.
Restored function that was gone.  Fixed calling function to get cell width and diag -- it changed signature.  Many updated libparcasl.cmake with new src classes.
Code runs slower than before!  Stefan problem benchmark is broken: it's not producing the results I reported in the paper.
Introduced 'ulong' type, which is not defined on macOS C++ compiler --so I replaced it back to 'unsigned long' in my_p4est_biomolecules.h.
I'll leave changes as is to avoid compilation errors.  Don't delete the mass conservation branch as that one contains the code that works as provided in the paper.