Resolved bug in interpolation caused by
a very rare case that was not fully handled in my_p4est_hierarchy. 

I modified one of the logical statements slightly which determines which tree to slide to alleviate this issue. 

Issue presents when a point for interpolation happens to fall *exactly* on the vertex of a quadrant, and therefore have the exact same coordinate as the P4EST_ROOT_LEN