Modified how we interface with the NS solver
So that we control the creation/destruction of the velocities rather than the NS solver controlling them and just giving us copies. 

This also eliminates a double interp that was happening -- we were interpolating vn_nodes to the new grid and the NS solver was doing it again. This might help speed things up a bit! 

Also added some nullify functionalities to NS so that we can nullify things that we have already destroyed before calling the NS destructor. 

Tested this with two grains merging and with coupled convergence test and it looks okay! Need to run valgrind though.