Made adjustments to ordering of save state, and grids to save/load
in order to be more consistent with the actual solution procedure -- a.k.a. to make sure we are not doing a sort of repeat calculation of the load state bc of how things are ordered. 

Now the state is saved at the end of the step (after the grid is advanced), which is consistent with where the load state would pick up (At the very beginning of the timestep)