Fixed awful bug that was throwing off results for hyp-paraboloid.  Back to single network approach.
Also added option for uniform random noise to testing interfaces: ellipsoid, paraboloid, and Gaussian.  Adapted the MongePatch class to hold its own h, which we can change to increase the resolution ONLY of the triangulated surface.  This was causing the problem.  Now, the h you use in the code and the h that the discretized Monge patch use ca be independent.