Cleaned up the dissolution BC for porous media
There was a slight mistake -- the Robin BC term coefficient was set to 1.0 for the porous media dissolution case, when it should have been set to Da. 

I've also reduced some code duplication for the Dirichlet term of the BC by just creating a constant function that returns the expression.