Fixed a bug I got in poisson nodes mls coming from fxn in utils
The function "find_closest_interface_location" in the class "my_p4est_utils" only worked properly on the assumption that the provided "dist" and "d" variables were positive values. However, there is one case where a provided "d" will be negative -- when the given node is a wall node, and the neighbor is a 2nd degree neighbor (with a negative distance, to note this). 

I modified the function to account for the fact that "d" and "dist" could be negative values. 

I also made some comments in poisson_nodes_mls as I was combing through anyways, and I did leave some residual debugging print statements (commented out) and a try/catch case in poisson nodes mls just in case the bug arises again (although it shouldn't). These can eventually be removed. 

I also left a mechanism for printing node neighbors in my own class for now, since Rochi and I are using it to investigate how 2 different neighbor functions operate. Will remove that later.