Deactivated the "2nd pass" of regularize front
Because it was sometimes creating tiny pools of solid between merging grains, which was crashing the NS solver. 

This part of the algorithm was already marked as optional by Daniil, so I'm just gonna bypass it to avoid the trouble it seems to create. I will have to keep an eye on this though