Fixed NS save/load
by utilizing the NS class' actual save/load functionality, that way we also save and load the local hodge and etc., which allows the loaded NS solver to converge immediately on load 

- I checked this load with a memory leak and it was clean, so that's good! 

However, it may have an eventual not totally clean exit from the time loop. I'll keep an eye out for that.