Resolving the "process_incoming_query" bug
This commit essentially contains some changes I made to both troubleshoot and (sketchily) resolve the process_incoming_query bug that was popping up especially in some of the porous media runs. 

I'm leaving some infrastructure in place that I added which is by default deactivated, but when this bug is triggered it will activate and therefore provide the user with a lot more information about the specifics of the crash. 

I've also left some notes about the fix I implemented, details about what the problem actually was, and some notes on how someone could reproduce it in the future if needed.